+++ a4kSubtitles/services/bsplayer.py
@@ -126,6 +126,9 @@ def build_search_requests(core, service_name, meta):
 
     action = 'searchSubtitles'
 
+    lang_ids = core.utils.get_lang_ids(meta.languages)
+    core.services[service_name].context.lang_ids = lang_ids
+
     params = (
         '<handle>{token}</handle>'
         '<movieHash>{filehash}</movieHash>'
@@ -136,7 +139,7 @@ def build_search_requests(core, service_name, meta):
         token=token,
         filesize=meta.filesize if meta.filesize else '0',
         filehash=meta.filehash if meta.filehash else '0',
-        lang_ids=','.join(core.utils.get_lang_ids(meta.languages)),
+        lang_ids=','.join(lang_ids),
         imdb_id=meta.imdb_id[2:]
     )
 
@@ -157,7 +160,7 @@ def parse_search_response(core, service_name, meta, response):
     if not results:
         return []
 
-    lang_ids = core.utils.get_lang_ids(meta.languages)
+    lang_ids = core.services[service_name].context.lang_ids
 
     def map_result(result):
         name = result.find('subName').text
@@ -175,7 +178,7 @@ def parse_search_response(core, service_name, meta, response):
             'lang': lang,
             'name': name,
             'rating': int(round(float(rating) / 2)) if rating else 0,
-            'lang_code': core.kodi.xbmc.convertLanguage(lang, core.kodi.xbmc.ISO_639_1),
+            'lang_code': core.utils.get_lang_id(lang, core.kodi.xbmc.ISO_639_1),
             'sync': 'true' if meta.filehash else 'false',
             'impaired': 'false',
             'color': 'gold',